# Use SequenceEqual in BigInteger.Equals