@huoyaoyuan 2023-09-01
Unfortunately `SequenceCompareTo` is in the reversed order and can't be used in `CompareTo`.